requirements.txt
praw
python-dotenv
spacy